% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{cadmium1}
\alias{cadmium1}
\title{Reproduction and survival data sets of chronic laboratory toxicity
tests of cadmium with Daphnia}
\usage{
data(cadmium1)
}
\description{
\itemize{
\item \strong{cadmium1}: Reproduction and survival data sets of chronic laboratory toxicity tests with
\emph{Daphnia magna} freshwater invertebrate exposed to five concentrations
of cadmium during 21 days. Five concentrations were
tested, with four replicates per concentration. Each replicate contained 10
organisms. Reproduction and survival were monitored at 10 time points.
}
}
\references{
Billoir, E., Delhaye, H., Forfait, C., Clement, B.,
Triffault-Bouchet, G., Charles, S. and Delignette-Muller, M.L. (2012)
Comparison of toxicity tests with different exposure time patterns: The added
value of dynamic modelling in predictive ecotoxicology, \emph{Ecotoxicology
and Environmental Safety}, 75, 80-86.
}
\keyword{datasets}
