% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{cadmium_daphnia}
\alias{cadmium_daphnia}
\title{Continuous response of toxicity test of cadmium on Daphnia}
\usage{
data(cadmium_daphnia)
}
\description{
\itemize{
\item \strong{cadmium_daphnia}: Exposure of \emph{Daphnia magna} to cadmium (5
concentrations including the
control) during 21 days. 10
time-points and 4 replicates of 10 animals. Length
data is collected (expressed in mm).
}

@references Billoir, E., Delignette-Muller, M.L., Pery, A.R.R. and
Charles, S. (2008) A Bayesian Approach to Analyzing Ecotoxicological Data,
\emph{Environmental Science & Technology}, 42 (23), 8978-8984.
}
\keyword{datasets}
