% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{chlordan}
\alias{chlordan}
\title{Reproduction and survival data sets of chronic laboratory toxicity
tests of chlordan with Daphnia}
\usage{
data(chlordan)
}
\description{
\itemize{
\item \strong{chlordan}: Reproduction and survival data sets of chronic laboratory toxicity tests with
\emph{Daphnia magna} freshwater invertebrate exposed to six concentrations
of one organochlorine insecticide (chlordan) during 21 days. Six concentrations were
tested, with 10 replicates per concentration. Each replicate contained one
organism. Reproduction and survival were monitored at 22 time points.
See Manar et al. (2009).
}
}
\references{
Manar, R., Bessi, H. and Vasseur, P. (2009) Reproductive effects
and bioaccumulation of chlordan in Daphnia magna, \emph{Environmental
Toxicology and Chemistry}, 28(10), 2150-2159.
}
\keyword{datasets}
