% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{chlordan_daphnia}
\alias{chlordan_daphnia}
\title{Continuous response of toxicity test of chlordan on Daphnia}
\usage{
data(chlordan_daphnia)
}
\description{
\itemize{
\item \strong{chlordan_daphnia}: Exposure of Daphnia magna to chlordan (6 concentrations including the
control) at day 21. Length data is collected
(expressed in mm).
See Manar et al. (2009)
}

@references Manar, R., Bessi, H. and Vasseur, P. (2009) Reproductive effects
and bioaccumulation of chlordan in Daphnia magna, \emph{Environmental
Toxicology and Chemistry}, 28(10), 2150-2159.
}
\keyword{datasets}
