% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{copper}
\alias{copper}
\title{Reproduction and survival data sets of chronic laboratory toxicity
tests of copper on Daphnia}
\usage{
data(copper)
}
\description{
\itemize{
\item \strong{copper}: Reproduction and survival data sets of chronic laboratory toxicity tests with
\emph{Daphnia magna} freshwater invertebrate exposed to five concentrations
of copper during 21 days. Five concentrations were
tested, with three replicates per concentration. Each replicate contained 20
organisms. Reproduction and survival were monitored at 16 time points.
}
}
\references{
Billoir, E., Delignette-Muller, M.L., Pery, A.R.R. and
Charles, S. (2008) A Bayesian Approach to Analyzing Ecotoxicological Data,
\emph{Environmental Science & Technology}, 42 (23), 8978-8984.
}
\keyword{datasets}
