% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dichromate}
\alias{dichromate}
\title{Survival data set of chronic laboratory toxicity tests of dichromate
with Daphnia}
\usage{
data(dichromate)
}
\description{
\itemize{
\item \strong{dichromate}: Survival data set of chronic laboratory toxicity tests with
\emph{Daphnia magna} freshwater invertebrate exposed to six concentrations
of one oxidizing agent (potassium dichromate) during 21 days. Six
concentrations were tested with one replicate of 50 organisms per concentration.
Survival is monitored at 10 time points.
}
}
\references{
Bedaux, J., Kooijman, SALM (1994) Statistical analysis of toxicity tests,
based on hazard modeling, \emph{Environmental and Ecological Statistics}, 1,
303-314.
}
\keyword{datasets}
