% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseResponse.R
\name{DoseResponse}
\alias{DoseResponse}
\alias{doseResponse}
\alias{doseResponse.BinaryData}
\alias{doseResponse.CountData}
\alias{doseResponse.ContinuousData}
\title{Prepare data for Dose-Reponse}
\usage{
doseResponse(data, ...)

\method{doseResponse}{BinaryData}(data, target.time = NULL, pool.replicate = FALSE, ...)

\method{doseResponse}{CountData}(data, target.time = NULL, pool.replicate = FALSE, ...)

\method{doseResponse}{ContinuousData}(data, target.time = NULL, pool.replicate = FALSE, ...)
}
\arguments{
\item{data}{an object used to select a method \code{doseResponse}}

\item{\dots}{Further arguments to be passed to generic methods}

\item{target.time}{Numeric. Default is \code{NULL}. By default, the last time
is considerd as the target-time. Otherwise, the argument set the target time.}

\item{pool.replicate}{Binary. Default is \code{FALSE}. IF \code{TRUE}, data
are summed according to groupd of the same \code{time} and \code{conc}.}
}
\value{
an object of class \code{DoseResponse}
}
\description{
Reshape data for using function on Dose-Reponse data#'
}
