% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_sim}
\alias{extract_sim}
\title{Extract simulation from the fit}
\usage{
extract_sim(fit)
}
\arguments{
\item{fit}{object of class \code{FitTT}}
}
\value{
return a \code{list} with 2 \code{data.frame}:
\itemize{
\item \code{mcmc}: a data.frame with simulated data.
\item \code{quantile}: a data.frame with quantile of the simulated data.
}
}
\description{
Extract simulation from the fit
}
