% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{plant10}
\alias{plant10}
\title{Continuous response of toxicity test on plants}
\usage{
data(plant10)
}
\description{
\itemize{
\item \strong{plant}: Plant species exposed to a given product during 21 days for the
vegetative vigour test. Shoot dry weight data is collected
}
}
\references{
Charles, S., Wu, D., Ducrot, V. (2021). How to account for
the uncertainty from standard toxicity tests in species sensitivity
distributions: An example in non-target plants. \emph{PLOS ONE}
}
\keyword{datasets}
