% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-XCX.R
\name{PlotRiskMeasure}
\alias{PlotRiskMeasure}
\alias{plot.XCX}
\title{Plot an object \code{XCX}}
\usage{
\method{plot}{XCX}(x, xlab = NULL, ylab = NULL, main = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{XCX}}

\item{xlab}{label of the x-axis}

\item{ylab}{label of the y-axis}

\item{main}{title of the graphic}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
a plot of class  \code{\link[ggplot2]{ggplot}}
}
\description{
Plot the median and 95\\% credible interval of the x\% Lethal or Effect
Concentration.
}
