% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{PPC}
\alias{PPC}
\alias{ppc}
\alias{ppc.ContinuousFitTT}
\alias{ppc.CountFitTT}
\alias{ppc.BinaryFitTT}
\title{Posterior Predictive Check data.frame for \code{FitTT} objects}
\usage{
ppc(fit, ...)

\method{ppc}{ContinuousFitTT}(fit, ...)

\method{ppc}{CountFitTT}(fit, ...)

\method{ppc}{BinaryFitTT}(fit, ...)
}
\arguments{
\item{fit}{An object of class \code{FitTT}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
An object of class 'PPC'
}
\description{
Create the \code{PPC} object to be pass in plot function for plotting the
Posterior Predictive Check.#'
}
