% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-FitTT.R, R/predict-FitTTBinary.R,
%   R/predict-FitTTContinuous.R, R/predict-FitTTCount.R
\name{Predict}
\alias{Predict}
\alias{predict}
\alias{predict.BinaryFitTT}
\alias{predict.ContinuousFitTT}
\alias{predict.CountFitTT}
\title{Prediction base on \code{FitTT}}
\usage{
predict(fit, ...)

\method{predict}{BinaryFitTT}(fit, display.conc = NULL, ...)

\method{predict}{ContinuousFitTT}(fit, display.conc = NULL, ...)

\method{predict}{CountFitTT}(fit, display.conc = NULL, ...)
}
\arguments{
\item{fit}{object of class \code{BinaryFitTT}, \code{CountFitTT} or
\code{ConitnuousFitTT}.}

\item{\dots}{Further arguments to be passed to generic methods}

\item{display.conc}{vector of concentrations values (x axis)}
}
\value{
A list
of 3 elements: 'display.conc' a vector of the display concentration of
exposure, 'mcmc': a data.frame of the mcmc and 'quantile' a data.frame
of quantiles.
}
\description{
Prediction base on \code{FitTT}
}
