% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-PriorPosterior.R, R/priorPosterior.R
\name{PriorPosterior}
\alias{PriorPosterior}
\alias{plot.PriorPosterior}
\alias{priorPosterior}
\alias{priorPosterior.BinaryFitTT}
\alias{priorPosterior.CountFitTT}
\alias{priorPosterior.ContinuousFitTT}
\alias{posterior}
\alias{posterior.FitTT}
\title{Return Prior and Posterior density of parameters of \code{FitTT} object}
\usage{
\method{plot}{PriorPosterior}(x, xlab = "value", ylab = NULL, main = NULL, ...)

priorPosterior(fit, size_sample, ...)

\method{priorPosterior}{BinaryFitTT}(fit, size_sample = 1000, ...)

\method{priorPosterior}{CountFitTT}(fit, size_sample = 1000, ...)

\method{priorPosterior}{ContinuousFitTT}(fit, size_sample = 1000, ...)

posterior(fit, ...)

\method{posterior}{FitTT}(fit, ...)
}
\arguments{
\item{x}{an object of class \code{PriorPosterior}}

\item{xlab}{label of the x-axis}

\item{ylab}{label of the y-axis}

\item{main}{title of the graphic}

\item{\dots}{Further arguments to be passed to generic methods}

\item{fit}{An object of class \code{FitTT}}

\item{size_sample}{graphical backend, can be \code{'generic'} or \code{'ggplot'}}
}
\value{
a data frame of class \code{PriorPosterior}

Return an object of class \code{Posterior}
}
\description{
Return Prior and Posterior density of parameters of \code{FitTT} object

Extract posterior of parameters from a \code{FitTT} object#'
}
