% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{propiconazole}
\alias{propiconazole}
\title{Survival data set of chronic laboratory toxicity tests of propiconazole
with Gammarus pulex}
\usage{
data(propiconazole)
}
\description{
\itemize{
\item \strong{propiconazole}: Survival data set of chronic laboratory toxicity tests with
\emph{Gammarus pulex} freshwater invertebrate exposed to eight concentrations
of one fungicide (propiconazole) during four days. Eight
concentrations were tested with two replicates of 10 organisms per concentration.
Survival is monitored at five time points.
}
}
\references{
Nyman, A.-M., Schirmer, K., Ashauer, R., (2012) Toxicokinetic-toxicodynamic
modelling of survival of \emph{Gammarus pulex} in multiple pulse exposures to
propiconazole: model assumptions, calibration data requirements and predictive
power, \emph{Ecotoxicology}, (21), 1828-1840.
}
\keyword{datasets}
