% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCX.R
\name{RiskMeasure}
\alias{RiskMeasure}
\alias{xcx}
\alias{xcx.FitTT}
\title{Predict X\% Concentration at the target time (default).}
\usage{
xcx(fit, x, ...)

\method{xcx}{FitTT}(fit, x, ...)
}
\arguments{
\item{fit}{An object used to select a method}

\item{x}{vector of values of LCx or ECX}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
returns an object of class \code{XCX} consisting in a data.frame with
the estimated ECx or LCx and their CIs
95\% (3 columns (values, CIinf, CIsup) and length(x) rows)
}
\description{
Predict median and 95\\% credible interval of the x\% Lethal or Effect
Concentration.
}
