% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{subst01_lymnaea}
\alias{subst01_lymnaea}
\title{Continuous response of toxicity test on snails}
\usage{
data(subst01_lymnaea)
}
\description{
\itemize{
\item \strong{subst01_lymnaea}: Exposure of snails to a given substance (6 concentrations including the
control) at day 56. Length of shell is collected
(expressed in mm).
}

@references Ducrot, V., Askem, C., Azam, D., Brettschneider, D., Brown,
R., Charles, S., Coke, M., Collinet, M., Delignette-Muller, M.L.,
Forfait-Dubuc, C., Holbech, H., Hutchinson, T., Jach, A., Kinnberg, K.L.,
Lacoste, C., Le Page, G., Matthiessen, P., Oehlmann, J., Rice, L.,
Roberts, E., Ruppert, K., Davis, J.E., Veauvy, C., Weltje, L., Wortham, R.
and Lagadic, L. (2014)
Development and validation of an OECD reproductive toxicity test guideline with
the pond snail Lymnaea stagnalis (Mollusca, Gastropoda),
\emph{Regulatory Toxicology and Pharmacology}, 70(3), 605-14.
}
\keyword{datasets}
