% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary-FitTT.R, R/summary-FitTTBinary.R,
%   R/summary-FitTTContinuous.R, R/summary-FitTTCount.R, R/summary-PPC.R
\name{Summary}
\alias{Summary}
\alias{summary.BinaryFitTT}
\alias{summary.ContinuousFitTT}
\alias{summary.CountFitTT}
\alias{summary.PPC}
\title{Summary of \code{FitTT} object}
\usage{
\method{summary}{BinaryFitTT}(object, quiet = FALSE, ...)

\method{summary}{ContinuousFitTT}(object, quiet = FALSE, ...)

\method{summary}{CountFitTT}(object, quiet = FALSE, ...)

\method{summary}{PPC}(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{FitTT}}

\item{quiet}{when \code{TRUE}, does not print}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
The function returns a list with the following information:
\item{Qpriors}{quantiles of the model priors}
\item{Qposteriors}{quantiles of the model posteriors}
\item{QLCx}{quantiles of LCx estimates}
}
\description{
This is the generic \code{summary} S3 method for the \code{FitTT} class.
It shows the quantiles of priors and posteriors on parameters and the quantiles
of the posteriors on the ECx and LCx estimates.

Summary function on \code{PPC} object
}
