% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{zinc}
\alias{zinc}
\title{Reproduction and survival data sets of a chronic laboratory toxicity
tests of zinc with Daphnia}
\usage{
data(zinc)
}
\description{
\itemize{
\item \strong{zinc}: Reproduction and survival data sets of a chronic laboratory toxicity tests with
\emph{Daphnia magna} freshwater invertebrate exposed to four concentrations
of zinc during 21 days. Four concentrations were
tested with three replicates per concentration. Each replicate contained 20
organisms. Reproduction and survival were monitored at 15 time points.
}
}
\references{
Billoir, E., Delignette-Muller, M.L., Pery, A.R.R. and
Charles, S. (2008) A Bayesian Approach to Analyzing Ecotoxicological Data,
\emph{Environmental Science & Technology}, 42 (23), 8978-8984.
}
\keyword{datasets}
