% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{zinc_daphnia}
\alias{zinc_daphnia}
\title{Continuous response of toxicity test of zinc on Daphnia}
\usage{
data(zinc_daphnia)
}
\description{
\itemize{
\item \strong{zinc_daphnia}: Exposure of Daphnia magna to zinc
(4 concentrations including the control)
during 21 days. 15 time-points and 3 replicates of 20
animals. Length data is collected (expressed in mm).
}
}
\references{
Billoir, E., Delignette-Muller, M.L., Pery, A.R.R. and
Charles, S. (2008) A Bayesian Approach to Analyzing Ecotoxicological Data,
\emph{Environmental Science & Technology}, 42 (23), 8978-8984.
}
\keyword{datasets}
